fix: window.MessageChannel not supported by JSDOM
