+++ packages/marko/src/runtime/queueMicrotask.js
@@ -1,0 +1,14 @@
+var promise;
+module.exports =
+  typeof queueMicrotask === "function"
+    ? queueMicrotask
+    : typeof Promise === "function" && (promise = Promise.resolve())
+    ? function(cb) {
+        promise.then(cb).catch(rethrow);
+      }
+    : setTimeout;
+function rethrow(err) {
+  setTimeout(function() {
+    throw err;
+  });
+}
